app/src/main/python/Audio_Parser/requirements.txt
praat-parselmouth==0.4.5
pandas==2.2.3
openpyxl==3.1.5